# tesea/tesea_responsive_production/api/openapi.yaml
openapi: 3.1.0
info:
  title: TESEA Academy Platform API
  version: 1.0.0
  description: Production contract boundary for the TESEA NECTA competence-based learning platform.
servers:
  - url: https://api.teseaacademy.ac.tz/api/v1
security:
  - bearerAuth: []
tags:
  - name: Auth
  - name: Users
  - name: Curriculum
  - name: Lessons
  - name: Questions
  - name: Assessments
  - name: Classes
  - name: Live lessons
  - name: Community
  - name: Moderation
  - name: Subscriptions
  - name: CRM
  - name: Support
  - name: Platform
paths:
  /auth/signup:
    post:
      tags: [Auth]
      security: []
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/SignupRequest'}}}}
      responses: {'201': {description: Account created, content: {application/json: {schema: {$ref: '#/components/schemas/AuthSession'}}}}, '409': {$ref: '#/components/responses/Conflict'}}
  /auth/signin:
    post:
      tags: [Auth]
      security: []
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/SigninRequest'}}}}
      responses: {'200': {description: Authenticated, content: {application/json: {schema: {$ref: '#/components/schemas/AuthSession'}}}}, '401': {$ref: '#/components/responses/Unauthorized'}}
  /auth/refresh:
    post:
      tags: [Auth]
      security: []
      responses: {'200': {description: Refreshed session, content: {application/json: {schema: {$ref: '#/components/schemas/AuthSession'}}}}}
  /users:
    get:
      tags: [Users]
      parameters: [{$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}]
      responses: {'200': {description: Users}}
    post:
      tags: [Users]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/UserWrite'}}}}
      responses: {'201': {description: User created}}
  /users/{userId}:
    get:
      tags: [Users]
      parameters: [{$ref: '#/components/parameters/UserId'}]
      responses: {'200': {description: User detail}, '404': {$ref: '#/components/responses/NotFound'}}
    patch:
      tags: [Users]
      parameters: [{$ref: '#/components/parameters/UserId'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/UserWrite'}}}}
      responses: {'200': {description: User updated}}
  /curriculum/subjects:
    get:
      tags: [Curriculum]
      responses: {'200': {description: Subject catalogue}}
  /curriculum/subjects/{subjectId}/forms/{formId}/tree:
    get:
      tags: [Curriculum]
      parameters: [{$ref: '#/components/parameters/SubjectId'}, {$ref: '#/components/parameters/FormId'}]
      responses: {'200': {description: Curriculum tree, content: {application/json: {schema: {$ref: '#/components/schemas/CurriculumTree'}}}}}
    put:
      tags: [Curriculum]
      parameters: [{$ref: '#/components/parameters/SubjectId'}, {$ref: '#/components/parameters/FormId'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CurriculumTree'}}}}
      responses: {'200': {description: Curriculum tree updated}}
  /lessons:
    get:
      tags: [Lessons]
      responses: {'200': {description: Lessons}}
    post:
      tags: [Lessons]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/LessonWrite'}}}}
      responses: {'201': {description: Lesson created}}
  /media/uploads:
    post:
      tags: [Lessons]
      description: Returns a signed upload target. Binary media is uploaded directly to object storage.
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [filename, contentType], properties: {filename: {type: string}, contentType: {type: string}, size: {type: integer}}}}}}
      responses: {'201': {description: Signed upload session}}
  /questions:
    get:
      tags: [Questions]
      responses: {'200': {description: Outcome-mapped question bank}}
    post:
      tags: [Questions]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/QuestionWrite'}}}}
      responses: {'201': {description: Question created}}
  /questions/import:
    post:
      tags: [Questions]
      description: Bulk ingestion for CSV, XLSX, JSON, QTI 2.1 and Moodle XML staging jobs.
      responses: {'202': {description: Import job accepted}}
  /questions/ai-generate:
    post:
      tags: [Questions]
      description: Creates AI-generated draft questions for tutor review; never auto-publishes.
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/AiQuestionRequest'}}}}
      responses: {'202': {description: Draft generation job accepted}}
  /assessments:
    get:
      tags: [Assessments]
      responses: {'200': {description: Assessments}}
    post:
      tags: [Assessments]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/AssessmentWrite'}}}}
      responses: {'201': {description: Assessment created}}
  /assessments/{assessmentId}/attempts:
    post:
      tags: [Assessments]
      parameters: [{$ref: '#/components/parameters/AssessmentId'}]
      responses: {'201': {description: Controlled attempt created}}
  /attempts/{attemptId}/submit:
    post:
      tags: [Assessments]
      parameters: [{name: attemptId, in: path, required: true, schema: {type: string}}]
      responses: {'200': {description: Submitted and outcome evidence queued}}
  /learners/{learnerId}/evidence:
    get:
      tags: [Assessments]
      parameters: [{name: learnerId, in: path, required: true, schema: {type: string}}]
      responses: {'200': {description: Outcome evidence and progression decisions}}
  /classes:
    get:
      tags: [Classes]
      responses: {'200': {description: Tutor classes}}
    post:
      tags: [Classes]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ClassWrite'}}}}
      responses: {'201': {description: Class created}}
  /classes/{classId}/roster/import:
    post:
      tags: [Classes]
      parameters: [{name: classId, in: path, required: true, schema: {type: string}}]
      responses: {'202': {description: Roster import job accepted}}
  /live-sessions:
    post:
      tags: [Live lessons]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/LiveSessionWrite'}}}}
      responses: {'201': {description: Live session scheduled}}
  /integrations/meetings/{provider}/connect:
    post:
      tags: [Live lessons]
      parameters: [{name: provider, in: path, required: true, schema: {enum: [zoom, teams, google-meet]}}]
      responses: {'200': {description: OAuth connection started or refreshed}}
  /community/posts:
    get:
      tags: [Community]
      responses: {'200': {description: Visible community posts}}
    post:
      tags: [Community]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CommunityPostWrite'}}}}
      responses: {'201': {description: Post created after moderation screening}}
  /community/posts/{postId}/report:
    post:
      tags: [Community]
      parameters: [{$ref: '#/components/parameters/PostId'}]
      responses: {'201': {description: Moderation case created}}
  /moderation/cases:
    get:
      tags: [Moderation]
      responses: {'200': {description: AI flags, tutor escalations and community reports}}
  /moderation/posts/{postId}/hide:
    post:
      tags: [Moderation]
      parameters: [{$ref: '#/components/parameters/PostId'}]
      responses: {'200': {description: Post hidden and audit event emitted}}
  /moderation/posts/{postId}/restore:
    post:
      tags: [Moderation]
      parameters: [{$ref: '#/components/parameters/PostId'}]
      responses: {'200': {description: Post restored}}
  /moderation/posts/{postId}:
    delete:
      tags: [Moderation]
      parameters: [{$ref: '#/components/parameters/PostId'}]
      responses: {'204': {description: Post removed from community visibility}}
  /subscriptions/me:
    get:
      tags: [Subscriptions]
      responses: {'200': {description: Current entitlement, billing validity and renewal state}}
  /subscriptions/me/invoices:
    get:
      tags: [Subscriptions]
      responses: {'200': {description: Billing history}}
  /subscriptions/checkout:
    post:
      tags: [Subscriptions]
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [planId], properties: {planId: {type: string}, returnUrl: {type: string, format: uri}}}}}}
      responses: {'201': {description: Provider checkout session}}
  /crm/leads:
    get:
      tags: [CRM]
      responses: {'200': {description: Lead pipeline}}
    post:
      tags: [CRM]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/LeadWrite'}}}}
      responses: {'201': {description: Lead captured}}
  /crm/leads/{leadId}/convert:
    post:
      tags: [CRM]
      parameters: [{name: leadId, in: path, required: true, schema: {type: string}}]
      responses: {'200': {description: Lead converted to customer onboarding workflow}}
  /crm/campaigns:
    post:
      tags: [CRM]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CampaignWrite'}}}}
      responses: {'201': {description: Campaign created}}
  /support/cases:
    get:
      tags: [Support]
      responses: {'200': {description: Support queue}}
    post:
      tags: [Support]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/SupportCaseWrite'}}}}
      responses: {'201': {description: Support case created}}
  /platform/settings:
    get:
      tags: [Platform]
      responses: {'200': {description: Effective platform settings}}
    put:
      tags: [Platform]
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: true}}}}
      responses: {'200': {description: Settings updated and audited}}
  /announcements:
    get:
      tags: [Platform]
      responses: {'200': {description: Announcements visible to requester}}
    post:
      tags: [Platform]
      responses: {'201': {description: Announcement created}}
  /events:
    post:
      tags: [Platform]
      description: Internal event ingestion boundary for asynchronous workflow integration.
      responses: {'202': {description: Event accepted}}
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    Cursor: {name: cursor, in: query, schema: {type: string}}
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 25}}
    UserId: {name: userId, in: path, required: true, schema: {type: string}}
    SubjectId: {name: subjectId, in: path, required: true, schema: {type: string}}
    FormId: {name: formId, in: path, required: true, schema: {type: string}}
    AssessmentId: {name: assessmentId, in: path, required: true, schema: {type: string}}
    PostId: {name: postId, in: path, required: true, schema: {type: string}}
  responses:
    Unauthorized: {description: Authentication required}
    NotFound: {description: Resource not found}
    Conflict: {description: Resource conflict}
  schemas:
    SignupRequest:
      type: object
      required: [name, email, password]
      properties: {name: {type: string}, email: {type: string, format: email}, password: {type: string, minLength: 8}, form: {type: string}}
    SigninRequest:
      type: object
      required: [email, password]
      properties: {email: {type: string, format: email}, password: {type: string}}
    AuthSession:
      type: object
      required: [user, token]
      properties: {user: {$ref: '#/components/schemas/User'}, token: {type: string}, refreshToken: {type: string}}
    User:
      type: object
      required: [id, name, email, role]
      properties: {id: {type: string}, name: {type: string}, email: {type: string, format: email}, role: {enum: [learner, tutor, admin, crm]}, status: {type: string}, scope: {type: object, additionalProperties: true}}
    UserWrite:
      type: object
      properties: {name: {type: string}, email: {type: string, format: email}, role: {type: string}, status: {type: string}, scope: {type: object, additionalProperties: true}}
    CurriculumTree:
      type: object
      required: [subjectId, formId, topics]
      properties: {subjectId: {type: string}, formId: {type: string}, topics: {type: array, items: {type: object, additionalProperties: true}}}
    LessonWrite:
      type: object
      required: [title, format, outcomeIds]
      properties: {title: {type: string}, format: {enum: [interactive-text, video, audio, pdf, shorts, live]}, outcomeIds: {type: array, items: {type: string}}, status: {enum: [draft, review, published]}}
    QuestionWrite:
      type: object
      required: [stem, type, outcomeIds]
      properties: {stem: {type: string}, type: {type: string}, outcomeIds: {type: array, items: {type: string}}, marks: {type: number}, answer: {type: object, additionalProperties: true}, status: {enum: [draft, review, approved, retired]}}
    AiQuestionRequest:
      type: object
      required: [outcomeIds, count]
      properties: {outcomeIds: {type: array, items: {type: string}}, count: {type: integer, minimum: 1, maximum: 100}, difficultyMix: {type: object, additionalProperties: true}, questionTypes: {type: array, items: {type: string}}}
    AssessmentWrite:
      type: object
      required: [title, type, outcomeIds]
      properties: {title: {type: string}, type: {enum: [practice, assignment, quiz, topic-test, exam]}, outcomeIds: {type: array, items: {type: string}}, questionIds: {type: array, items: {type: string}}, evidenceWeight: {type: number}}
    ClassWrite:
      type: object
      required: [name, subjectId, formId]
      properties: {name: {type: string}, subjectId: {type: string}, formId: {type: string}, section: {type: string}, outcomeIds: {type: array, items: {type: string}}}
    LiveSessionWrite:
      type: object
      required: [classId, outcomeId, evidenceGoal, startsAt]
      properties: {classId: {type: string}, outcomeId: {type: string}, evidenceGoal: {type: string}, startsAt: {type: string, format: date-time}, provider: {enum: [zoom, teams, google-meet]}}
    CommunityPostWrite:
      type: object
      required: [channelId, body]
      properties: {channelId: {type: string}, outcomeId: {type: string}, body: {type: string, maxLength: 5000}}
    LeadWrite:
      type: object
      required: [name, type]
      properties: {name: {type: string}, type: {enum: [learner, parent, school, partner]}, email: {type: string, format: email}, phone: {type: string}, source: {type: string}, consent: {type: boolean}, nextAction: {type: string}}
    CampaignWrite:
      type: object
      required: [name, audience, content]
      properties: {name: {type: string}, audience: {type: object, additionalProperties: true}, content: {type: object, additionalProperties: true}, scheduleAt: {type: string, format: date-time}}
    SupportCaseWrite:
      type: object
      required: [requesterId, category, subject]
      properties: {requesterId: {type: string}, category: {type: string}, subject: {type: string}, body: {type: string}, priority: {enum: [low, normal, high, urgent]}, context: {type: object, additionalProperties: true}}